backend/requirements.txt
Django==3.0.6
django-debug-toolbar==1.11
pylint==2.3.1
pylint-django==2.0.7
pylint-json2html==0.1.0
anybadge==1.1.1
constantly==15.1.0
pep8==1.7.1
pycrypto==2.6.1
pytz==2019.3
djangorestframework==3.11.0
coverage==5.1
django-cors-headers==3.2.1
django-silk==4.0.1
